Statistic reporting is now fully configurable.
Refactored code to cleanly seperate status encoding and recording. This
most evident in the creation of an Encoder class followed by the creation
of a Recorder class where the encoder.status_encoder is passed as a
parameter.

Generalized the use of the Recorder class to be used for both tweets
collection and statistics logging. This include the configuration file
where the recorder and stats section have substantial overlap.

Added config_recorder function to simply the creation of various types
of recorders (steam, watched, rotating).

Added missing methods to the SerialListener.